Move TransactionService extensions back to its own file
TransactionService is UI component and deserves another name.
